Remove dependency on libpython for Linux
pyscard is linked against libpython. This is considered a bad habit in
Debian as it pulling extra dependencies when installing.

The python interpreter itself is already statically linked against
libpython, linking the extension to it is creating unneeded dependencies
and might increase the memory usage.

This change fixes Debian bug #689923
http://bugs.debian.org/689923