Fix a problem with Windows Remote Desktop
SCardGetStatusChange() fails to work if rgReaderStates[] is not cleared.

See https://msdn.microsoft.com/en-us/library/aa379773.aspx
" Important  Each member of each structure in this array must be
initialized to zero and then set to specific values as necessary. If
this is not done, the function will fail in situations that involve
remote card readers. "

Thanks to Magnus Svensson for the bug report and patch
http://sourceforge.net/p/pyscard/bugs/21/