Fix failing TestUtil.test_toGSM3_38Bytes()
Prior to this patch `TestUtil.test_toGSM3_38Bytes()` was calling
`toGSM3_38Bytes()` with the byte string b'@\xc3\xb9Pascal'. The ù was
encoded as 2 bytes.

This could not work, as the ù key of `__dic_GSM_3_38__` was encoded in
iso-8859-1, i.e. as a single byte.

By making all the keys 1-char unicode strings, and explicitly converting
byte-strings to unicode we enable `toGSM3_38Bytes()` to accept both
unicode and byte strings.