_Transmit: use SCARD_PCI_RAW for undefined protocol
If the connection to the card/reader was done using SCARD_SHARE_DIRECT
then the card protocol has not been negociated and is undefined.

This is not an error and _Transmit should not fail in this case.
For example this is used to read/write memory cards.

Thanks to Leif Johansson for the bug report.
"T0/T1 marshaling confusion"
https://github.com/LudovicRousseau/pyscard/issues/31