CardMonitoring: stop the looping only if PCSC exited
The fix in 37b6997d05ee9feae41ad53c0189474fd23fb3be was not correct. The
loop should be exited only when PC/SC is no more running (return code
SCARD_E_NO_SERVICE) but not in the other cases.

The fix above created the regression reported in
https://github.com/LudovicRousseau/pyscard/issues/64

When an exception occured (like card not responding) the loop was exited
and the program just wait forever.