Updates text to be tolerant of macOS typo.
This change is necessary to build on modern Macs
with nix. See https://github.com/NixOS/nixpkgs/issues/91716

" Describe the bug
The Python package pyscard version 1.9.9 doesn't build on Darwin.

To Reproduce
Steps to reproduce the behavior:
nix-env -iA pythonPackages.pyscard on Darwin

Expected behavior
pyscard is successfully installed

This seems to have been introduced with the update to version 1.9.9 of
the package. Previously, a patch was provided to work around a buggy
error message that's checked during tests of the module.

This test was refactored in the 1.9.9 release to better detect old
versions of macOS and OS X.

As such, the patch stopped applying. The patch was retained in the tree
but had its usage removed in this commit, in response to an issue opened
on pyscard. While this fixed the case for building on some platforms,
python from nix on modern macOS still emits a value that causes tests to
fail.

The following program demonstrates that python from nixpkgs will still
report a version that will cause the test to believe it needs to check
against the misspelling:

nix run nixpkgs.python -c python -c '
     from distutils.util import get_platform
     print get_platform()'
macosx-10.12-x86_64

However, modern macOS will emit the correctly-spelled message, causing
this test to fail. "