Handle removal of the last reader on Windows
On Windows, when the last reader is removed the PC/SC manager is
stopped. A PC/SC call will receive SCARD_E_SERVICE_STOPPED.

The idea here is to renew the PC/SC context in getReaderNames().

Thanks to Lemon Gong for the patch
"fix #74: plugout the last device will raise an smartcard.Exceptions #93"
https://github.com/LudovicRousseau/pyscard/pull/93/files

This change should also fix:
- "Plugout an PCSC device will raise ListReadersException(hresult) on windows #74"
  https://github.com/LudovicRousseau/pyscard/issues/74
- "Windows smartcard service is started only once #114"
  https://github.com/LudovicRousseau/pyscard/issues/114