PCSCCardRequest: fix active polling (0.1 sec loop)
This change mostly reverts 0738c43e902485583cfb6de91db8ed016f1a4d76 from
Jean-Daniel in 2007.

We use the PnP special reader "\\?PnP?\Notification" to be notified when
a new reader is connected.

We also use a real timout in SCardGetStatusChange() of 5 seconds
(instead of 0 and a sleep of 0.1 sec) to avoid an active polling.
The call to SCardGetStatusChange() is canceled using SCardCancel() at
the end of the user defined timeout.