PCSCCardRequest: fix a bug when no reader is present
$ python3 PCSCCardRequest.py
Insert a new card within 10 seconds
Traceback (most recent call last):
  File "/../smartcard/pcsc/PCSCCardRequest.py", line 362, in <module>
    cs = cr.waitforcard()
         ^^^^^^^^^^^^^^^^
  File "/.../smartcard/pcsc/PCSCCardRequest.py", line 165, in waitforcard
    state[1] = state[1] & (0xFFFFFFFF ^ SCARD_STATE_CHANGED)
    ~~~~~^^^
TypeError: 'tuple' object does not support item assignment

The problem occurs when no reader is connected and
"\\?PnP?\Notification" is the only reader in the list.

The newstates returned by SCardGetStatusChange() is a list of tuple. So
the tuple cannot be modified.
In fact this code is useless since SCARD_STATE_PRESENT will never be set
for the PnP special reader or for a reader that has been removed.