Test readergroups: demonstrate pathological issues
This PR introduces extensive testing of the readergroups class, and
achieves 100% test coverage of the file.

However, readergroups has pathological implementation problems, which
are captured in the tests marked with xfail or that use pytest.raises()
to capture failures:

- Calling readergroups.addreadergroup() causes a RecursionError
- Calling readergroups.removereadergroup() causes a ValueError
- Calling readergroups.getreadergroups() returns a hard-coded empty list

There is at least one side effect I've found when readergroups is
inherited by the PCSC reader groups subclass, but I haven't added those
tests yet.

Now that the problems are captured in the test suite, I intend to work
to remove the readergroups class entirely.