PCSCCardRequest: use a local PC/SC context to avoid locks
SCardGetStatusChange() should not be called from the same PC/SC context
used by all the other PC/SC functions. Because of a mutex in pcsc-lite
the calls using the same context are exclusive. So the performances were
bad.

The problem was discovered by the framework testsuite that was running
very slow on GNU/Linux (but not on Windows or macOS).