waitforcardevent(): do not miss events between 2 calls
We now use the states returned by the previous SCardGetStatusChange()
call as initial state for the next call.
Since pcsc-lite includes and use the number of card events in the high
word of dwEventState the function SCardGetStatusChange() returns if an
event was missed.

Thanks to Lars Lengersdorf for the bug report
" Deadlock with multiple readers and quick reads #220 "
https://github.com/LudovicRousseau/pyscard/issues/220